feat: likelihood without problog
Now the computation of the likelihood is done without using ProbLog. Inference in ProbLog is
still used to calculate the 'count' column of the configurations tables.

A new column named 'active_rules' was added to the configurations tables. It stores which
rules are active for the given configurations, and makes it easier to compute the associated
likelihoods. It will be useful for computing the exact maximization of the parameters also.