removed inconsistent catch, learning info added
The error ProbLog was raising because of inconsistent evidence corresponds to cases where
the observation in the dataset is not compatible with the given structure. A description of why
the error should have occurred was provided also.

Learning information of each EM cycle started to be gathered by a function named
self._update_learning_info(). For each cycle it stores the parameters of each rule, the time the
cycle ended and the Log-Likelihood for that cycle. At the end of the learning function it
returns a DataFrame where each row is one cycle.