calculation of initial log_likelihood
Changes in learner.py and calculations.py to calculate and log the initial value of the expected
log-likelihood, before the end of the first EM cycle is reached. A new function log-likelihood
in calculations.py was created to calculate the likelihood for the whole model (and not just one
head).