feat: suppress evidences in problog string
Function parser.build_problog_model_str() had been updated to allow suppressing evidences when
creating the model string. This is useful when we are sampling instead of using the evaluator,
because we can simply the evidences at the end instead of searching where they are in the
string. When we are sampling, there is no need to pre-define the variables which are evidence.