managing for prod version after relational changes
Changes in parser.py made in the commit that handled relational data
weren't compatible with production version of the propositional learner.
New changes were made to handle this. This version can be considered a
production version of a propositional datasets learner. A new branch
named develop will be created, where the development of the relational
learner will take place.